+++ apps/server/src/shared/repo/index.ts
@@ -11,3 +11,5 @@ export * from './files';
 export * from './lesson';
 export * from './news';
 export * from './storageprovider';
+export * from './submission';
+export * from './task';